Read successfully RG 18-210 in rdflib.js demo
The demo is very primitive at this point, but it works.

Also, reading directly from a URL is not implemented yet -
I suspect I am going to need to implement Solid
for this:

https://github.com/SolidOS/solid-tutorial-rdflib.js

Alternatively, I could store the RDFs
in a hiddent input field. However, this would not
work with a number of RDFs.
Ultimately, I need to implement reading
the entire gbad directory structure.